rewrite CompoundColumnType to render sub-columns
Instead of trying to "glue" all the sub-columns together they are
always rendered and there CellView is passed to the "compound_column"
CellView as the value property.

The CellView of each sub-column must rendered as any other
but with attributes['compound'] set to true to inform it's a sub cell.