Make Datagrid configuration immutable
To reduce the size the Datagrid interface, and keep things simple,
changing the columns on a datagrid is no longer possible and
the event-dispatcher for updating a DatagridView is replaced by
a single viewBuilder callable.

Setting the data on a datagrid is now possible only once!
As there should not be a need to change this after the view is created.