Provide the template block_names by BaseType
This logic was already performed in the Twig DatagridRender, but keeping it in the BaseType
reduces duplicating logic and allows for greater flexibility.

Note that child columns inherit the name of there parent column (which was the main reason for CompoundColumn refactoring,
as they first were not aware of there parent). But not that a child-column doesn’t have direct access to the view of the parent.