Introduce a better way of reusing datagrids
The DatagridFactory::createDatagrid() now tries to find the Datagrid “configuration” from a registry, rather then creating a Datagrid manually.

This allows to create a Datagrid configurator that can be reused multiple times, similar to Column types, but less advanced.
The main bonus, you only need the DatagridFactory in a Controller to create a new Datagrid rather then various dependencies.